Export store creator to separate module
