+++ app/react/store/configure-store.js
@@ -1,0 +1,21 @@
+import { createStore, applyMiddleware, compose } from 'redux';
+import thunkMiddleware from 'redux-thunk';
+import rootReducer from '../reducers/index';
+
+const applyDevTools = function applyDevTools() {
+  if (process.env.NODE_ENV === 'development' && window.devToolsExtension) {
+    return window.devToolsExtension();
+  }
+  return f => f;
+};
+
+export default function configureStore(initialState) {
+  return createStore(
+    rootReducer,
+    initialState,
+    compose(
+      applyMiddleware(thunkMiddleware),
+      applyDevTools(),
+    )
+  );
+}
